(1248) Split up statsModel
