+++ packages/api/server/models/statsModel/numberOfActiveUsers.js
@@ -1,0 +1,20 @@
+const { sequelize } = require('#db/models');
+
+module.exports = async () => {
+    const rows = await sequelize.query(
+        `SELECT
+            COUNT(*) AS total
+        FROM users
+        LEFT JOIN localized_organizations AS organizations ON users.fk_organization = organizations.organization_id
+        WHERE
+            users.fk_status='active'
+            AND
+            organizations.active = TRUE
+        `,
+        {
+            type: sequelize.QueryTypes.SELECT,
+        },
+    );
+
+    return rows[0].total;
+};
