+++ packages/api/server/models/statsModel/numberOfPlans.js
@@ -1,0 +1,18 @@
+const { sequelize } = require('#db/models');
+
+module.exports = async (departement) => {
+    const rows = await sequelize.query(
+        `
+        SELECT COUNT(*) AS total 
+        FROM plans2
+        LEFT JOIN plan_departements on plan_departements.fk_plan = plans2.plan_id
+        WHERE closed_at IS NULL
+        ${departement ? `AND fk_departement = '${departement}'` : ''}
+        `,
+        {
+            type: sequelize.QueryTypes.SELECT,
+        },
+    );
+
+    return rows[0].total;
+};
