+++ packages/api/server/models/statsModel/numberOfReviewedComments.js
@@ -1,0 +1,1 @@
+module.exports = () => Promise.resolve(-1);
