+++ packages/api/server/models/statsModel/numberOfUsers.js
@@ -1,0 +1,17 @@
+const { sequelize } = require('#db/models');
+
+module.exports = async (departement) => {
+    const rows = await sequelize.query(
+        `
+        SELECT COUNT(*) from users
+        LEFT JOIN localized_organizations on users.fk_organization = localized_organizations.organization_id
+        WHERE fk_status = 'active'
+        ${departement ? `AND departement_code = '${departement}'` : ''}
+        `,
+        {
+            type: sequelize.QueryTypes.SELECT,
+        },
+    );
+
+    return rows[0].count;
+};
